Cleanup et1d rust solvers
