+++ elastic-tube-1d/solid-rust/src/main.rs
@@ -2,17 +2,7 @@ use precice;
 use std::env;
 use std::process::ExitCode;
 
-fn solid_compute_solution(pressure: &[f64], cross_section_length: &mut [f64]) {
-    assert!(pressure.len() == cross_section_length.len());
-    const E: f64 = 10000.0;
-    const C_MK2: f64 = E / std::f64::consts::FRAC_2_SQRT_PI;
-    const PRESSURE0: f64 = 0.0;
-    let new: Vec<_> = pressure
-        .iter()
-        .map(|p| ((PRESSURE0 - 2.0 * C_MK2) / (p - 2.0 * C_MK2)).powi(2))
-        .collect();
-    cross_section_length.copy_from_slice(&new[..]);
-}
+mod solver;
 
 fn main() -> ExitCode {
     println!("Starting Solid Solver...");
@@ -45,7 +35,7 @@ fn main() -> ExitCode {
     let grid_size = CHUNK_SIZE * dimensions as usize;
     let grid: Vec<f64> = {
         let mut v: Vec<f64> = vec![0_f64; grid_size];
-        const DX : f64 = TUBE_LENGTH / DOMAIN_SIZE as f64;
+        const DX: f64 = TUBE_LENGTH / DOMAIN_SIZE as f64;
         for i in 0..CHUNK_SIZE - 1 {
             v[i * dimensions as usize] = DX * i as f64;
         }
@@ -87,7 +77,7 @@ fn main() -> ExitCode {
             &mut pressure[..],
         );
 
-        solid_compute_solution(&pressure, &mut cross_section_length);
+        solver::solid_compute_solution(&pressure, &mut cross_section_length);
 
         participant.write_data(
             mesh_name,
